refactor : page 하위 클래스 util, common 패키지로 분리 (#559)
