+++ src/main/java/com/clubber/ClubberServer/domain/notice/service/NoticeService.java
@@ -5,10 +5,9 @@ import com.clubber.ClubberServer.domain.notice.dto.GetNoticeResponse;
 import com.clubber.ClubberServer.domain.notice.dto.GetNoticesAtMainResponse;
 import com.clubber.ClubberServer.domain.notice.exception.NoticeNotFoundException;
 import com.clubber.ClubberServer.domain.notice.repository.NoticeRepository;
-import com.clubber.ClubberServer.global.page.PageResponse;
+import com.clubber.ClubberServer.global.common.page.PageResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
